refactor create teams (#190)
* refactor create teams

* Update loading messagge
