+++ src/app/page.js
@@ -7,7 +7,9 @@ import { PulseLoader } from 'react-spinners'
 
 export default function Home() {
   const [isLoading, setIsLoading] = useState(true)
-  const [loadingMessage, setLoadingMessage] = useState('Loading Create Teams') // Always start with same message for SSR
+  const [loadingMessage, setLoadingMessage] = useState(
+    'Loading Loons Team Balancer'
+  ) // Always start with same message for SSR
   const router = useRouter()
 
   // Update loading message after component mounts (client-side only)
@@ -20,7 +22,6 @@ export default function Home() {
     if (lastAccess && parseInt(lastAccess) >= fiveMinutesAgo) {
       setLoadingMessage('Checking authentication') // Server probably warm
     }
-    // Otherwise keep 'Loading Create Teams' for likely cold start
   }, [])
 
   useEffect(() => {
